requirements.txt
fastapi
uvicorn

SQLAlchemy
psycopg2-binary
alembic

python-dotenv